backend/pyproject.toml

[project]
name = "mirofish-v2-backend"
version = "0.1.0"
description = "MiroFish V2 Backend"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
    "pydantic",
    "hindsight-client>=0.5.1",
    "websockets>=16.0",
    "pypdf>=6.10.2",
    "duckduckgo-search>=8.1.1",
    "python-multipart>=0.0.26",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
